refactor(backend): move PATCH story endpoint to story router, remove StoryService leak from projects router
