+++ backend/tests/story_engine/test_projects_api.py
@@ -901,122 +901,3 @@ async def test_my_project_user_isolation(
         text('DELETE FROM "user" WHERE id = :id'), {"id": user_b.id}
     )
     await db_session.commit()
-
-
-# ---------------------------------------------------------------------------
-# PATCH /projects/{project_id}/story/{story_id} — update story meta/identity
-# ---------------------------------------------------------------------------
-
-
-def _patch_story_url(project_id: uuid.UUID, story_id: uuid.UUID) -> str:
-    return f"/api/comic-builder/v2/projects/{project_id}/story/{story_id}"
-
-
-async def test_patch_story_updates_meta(
-    api_client: AsyncClient,
-    user: User,
-    project: Project,
-    story: Story,
-) -> None:
-    """Sending a new meta object replaces the stored meta."""
-    new_meta = {"tone": "Comedic", "comicStyle": "Manga"}
-    response = await api_client.patch(
-        _patch_story_url(project.id, story.id),
-        headers=_auth_headers(user.id),
-        json={"meta": new_meta},
-    )
-
-    assert response.status_code == 200
-    assert response.json()["meta"] == new_meta
-
-
-async def test_patch_story_updates_name(
-    api_client: AsyncClient,
-    user: User,
-    project: Project,
-    story: Story,
-) -> None:
-    """Sending a new name updates it while leaving description untouched."""
-    original_description = story.description
-    new_name = "A Brand New Title"
-    response = await api_client.patch(
-        _patch_story_url(project.id, story.id),
-        headers=_auth_headers(user.id),
-        json={"name": new_name},
-    )
-
-    assert response.status_code == 200
-    body = response.json()
-    assert body["name"] == new_name
-    assert body["description"] == original_description
-
-
-async def test_patch_story_updates_description(
-    api_client: AsyncClient,
-    user: User,
-    project: Project,
-    story: Story,
-) -> None:
-    """Sending a new description updates it while leaving name untouched."""
-    original_name = story.name
-    new_description = "A completely revised one-liner for this comic."
-    response = await api_client.patch(
-        _patch_story_url(project.id, story.id),
-        headers=_auth_headers(user.id),
-        json={"description": new_description},
-    )
-
-    assert response.status_code == 200
-    body = response.json()
-    assert body["description"] == new_description
-    assert body["name"] == original_name
-
-
-async def test_patch_story_updates_all_fields(
-    api_client: AsyncClient,
-    user: User,
-    project: Project,
-    story: Story,
-) -> None:
-    """All three fields can be updated in a single request."""
-    new_meta = {"tone": "Dark", "comicStyle": "Noir"}
-    new_name = "Shadows of the Cart"
-    new_description = "A noir retelling of Chekhov set in rural Russia."
-    response = await api_client.patch(
-        _patch_story_url(project.id, story.id),
-        headers=_auth_headers(user.id),
-        json={"meta": new_meta, "name": new_name, "description": new_description},
-    )
-
-    assert response.status_code == 200
-    body = response.json()
-    assert body["meta"] == new_meta
-    assert body["name"] == new_name
-    assert body["description"] == new_description
-
-
-async def test_patch_story_404_unknown_story(
-    api_client: AsyncClient,
-    user: User,
-    project: Project,
-) -> None:
-    """Patching a non-existent story returns 404."""
-    response = await api_client.patch(
-        _patch_story_url(project.id, uuid.uuid4()),
-        headers=_auth_headers(user.id),
-        json={"name": "Ghost Story"},
-    )
-    assert response.status_code == 404
-
-
-async def test_patch_story_401_no_token(
-    api_client: AsyncClient,
-    project: Project,
-    story: Story,
-) -> None:
-    """PATCH without a token returns 401."""
-    response = await api_client.patch(
-        _patch_story_url(project.id, story.id),
-        json={"name": "Unauthorized"},
-    )
-    assert response.status_code == 401
